added the youtube player feature, resolved path issues ti the diet detail page and structured the code
