+++ lib/features/explore/domain/use_cases/get_plans_use_case.dart
@@ -1,5 +1,5 @@
 import '../../../plans/data/models/plan_model.dart';
-import '../../data/repositories/plan_repository.dart';
+import '../../../plans/data/repositories/plan_repository.dart';
 
 class GetPlansUseCase {
   final PlanRepository repository;
