+++ lib/features/plans/domain/use_cases/get_favorite_plans_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:yeetfit/features/explore/data/repositories/plan_repository.dart';
+import 'package:yeetfit/features/plans/data/repositories/plan_repository.dart';
 import 'package:yeetfit/features/plans/data/models/plan_model.dart';
 
 class GetFavoritePlansUseCase {
