+++ lib/features/plans/domain/use_cases/toggle_favorite_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:yeetfit/features/explore/data/repositories/plan_repository.dart';
+import 'package:yeetfit/features/plans/data/repositories/plan_repository.dart';
 
 class ToggleFavoriteUseCase {
   final PlanRepository repository;
